Add WIP progress on switching to MTL style like DinoRush
