+++ library/Obelisk.hs
@@ -1,64 +1,30 @@
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE GeneralisedNewtypeDeriving #-}
+
 module Obelisk where
 
-import Control.Monad.State
+import Obelisk.Homogenous
 
-import Control.Monad
+import Obelisk.Config
+import Obelisk.State
+
+import Control.Monad.Reader
+import Control.Monad.State
+import Control.Monad.Catch
+
 import Control.Concurrent (threadDelay)
-import Foreign.C.Types
+import Foreign.C.Types ( CInt )
 import qualified Data.Text as T
 import qualified SDL.Vect as SDL
 import qualified SDL
 import qualified SDL.Event as SDL
 import qualified SDL.Video.Renderer as SDL
 
-import Data.Bifunctor
+
 import Data.Word
 import Linear
 import Control.Lens
 import SDL.Primitive as SDL
-
-type HV2 = V3 --TODO propagate this to indicate homoCoordsgenous coordinates
-type Line = (HV2 CInt, HV2 CInt) -- After homoCoordsgenous coordinates its a v3
-
-
-zoomT :: (Num a) => a -> V3 (V3 a)
-zoomT scale = V3 (V3 scale 0     0)
-                (V3 0     scale 0)
-                (V3 0     0     1)
-
-translate :: (Num a) => a -> a -> V3 (V3 a)
-translate x y = V3 (V3 1 0 x)
-                   (V3 0 1 y)
-                   (V3 0 0 1)
-
-rotation :: Double -> V3 (V3 Double)
-rotation theta = V3 (V3 (cos theta) (-sin theta) 0)
-                    (V3 (sin theta) (cos theta)  0)
-                    (V3  0           0           1)
-
-rotateAround :: Double -> V2 Double -> V3 (V3 Double)
-rotateAround theta (V2 x y) = translate x y !*! rotation theta !*! translate (-x) (-y)
-
-type M22Affine t = V3 (V3 t) -- TODO use this type alias??
-
-idv3 = V3 (V3 1 0 0) (V3 0 1 0) (V3 0 0 1) :: V3 (V3 CInt)
-
-applyAffineTransform :: M22Affine CInt -> [Line] -> [Line]
-applyAffineTransform t = fmap (bimap (t !* ) (t !*))
-
-appDTFloor :: M22Affine Double -> [Line] -> [Line]
-appDTFloor t = fmap (bimap f f)
-    where
-        f = fmap floor . (t !*) . fmap fromIntegral :: V3 CInt -> V3 CInt
-        --Convert to doubles, apply the transform then floor it
-
-homoCoords :: (Num a) => V2 a -> HV2 a
-homoCoords (V2 x y) = V3 x y 1
-
---TODO make sure the HV2 usage is correct
-dropHomoCoords :: (Num a) => HV2 a -> V2 a
-dropHomoCoords (V3 x y _) = V2 x y
 
 --COLORS
 white = SDL.V4 maxBound maxBound maxBound maxBound :: Color
@@ -185,35 +151,12 @@ vectorAngle (V2 x y)
 
 (screenWidth, screenHeight) = (640, 480) :: (CInt,CInt)
 
-data WallType = EW | FW | DW --Empty Wall, Full Wall, Door Wall
-type WorldTiles = [[WallType]]
 
-rFW = repeat FW
-rEW = repeat EW
+
+
 --ACCESSED godBoltMap !! y !! x style
-godboltMap :: [[WallType]]
-godboltMap = [take 10 rFW,
-              FW : take 3 rEW ++ [FW] ++ take 4 rEW ++ [FW],
-              FW : take 3 rEW ++ [FW] ++ take 4 rEW ++ [FW],
-              take 3 rFW ++ [DW] ++ [FW] ++ take 4 rEW ++ [FW],
-              FW : take 3 rEW ++ [FW] ++ take 4 rEW ++ [FW],
-              [FW,DW] ++ take 3 rFW ++ take 4 rEW ++ [FW],
-              FW : take 3 rEW ++ take 4 rFW ++ [DW, FW],
-              FW : take 8 rEW ++ [FW],
-              FW : take 8 rEW ++ [FW],
-              take 10 rFW]
-
---In the style of https://github.com/jxv/diner/library/DinoRo-rush/blob/mastush/State.hs
-data PVars = PVars {
-                position :: V2 Double,
-                direction :: V2 Double
-             }
 
 
-data Vars = Vars {
-                player :: PVars,
-                map :: WorldTiles
-            }
 
 gameTick :: ScreenHandles -> StateT Vars IO ()
 gameTick hs = do
@@ -241,6 +184,7 @@ drawDebug (window, screenSurface, screenRenderer) gs = do
 --TODO make minimap rotate around the player
 --TODO mouse zoom handling
 
+--TODO REMOVE
 type ScreenHandles = (SDL.Window, SDL.Surface, SDL.Renderer)
 
 initVars = Vars p godboltMap
@@ -280,10 +224,21 @@ main = do
 
     let hs = (window, screenSurface, screenRenderer)
 
+    --TODO incorporate
+    let cfg = Config {
+                cWindow = window,
+                cRenderer = screenRenderer,
+                cSurface = screenSurface
+            }
+
     --CONSTS
     --TODO INIT VARS
+
     execStateT (mainLoop hs) initVars
 
     SDL.freeSurface screenSurface
     SDL.destroyWindow window
     SDL.quit
+
+newtype Obelisk a = Obelisk (ReaderT Config (StateT Vars IO) a)
+    deriving (Functor, Applicative, Monad, MonadReader Config, MonadState Vars, MonadIO, MonadThrow, MonadCatch)
