+++ obelisk.cabal
@@ -28,6 +28,7 @@ library
       , lens >= 5.0.1
       , directory >= 1.3.6.2
       , mtl >= 2.2.2
+      , exceptions >= 0.10.4
   default-language: Haskell2010
 
 executable obelisk-exe
